updated pom file and docker file
