+++ digitaldots-java8/function/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.digitaldots</groupId>
-    <artifactId>java8maven</artifactId>
+    <artifactId>function</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
