+++ digitaldots-java8/function/src/main/java/com/openfaas/function/Handler.java
@@ -1,0 +1,14 @@
+package com.openfaas.function;
+
+import com.digitaldots.function.BeverFunctionHandler;
+import com.openfaas.model.IRequest;
+import com.openfaas.model.IResponse;
+
+public class Handler extends com.openfaas.model.AbstractHandler{
+    
+    public IResponse Handle(IRequest request) {
+        BeverFunctionHandler beverHandler = new BeverFunctionHandler();
+        return beverHandler.Handle(request);
+    }
+}
+
