--- template/digitaldots-java8/function/src/main/java/com/digitaldots/function/BeverHandler.java
@@ -1,8 +1,0 @@
-package com.openfaas.function;
-
-import com.openfaas.model.IRequest;
-import com.openfaas.model.IResponse;
-
-public abstract class BeverHandler  extends com.openfaas.model.AbstractHandler{
-    public abstract IResponse Handle(IRequest var1);
-}
