--- template/digitaldots-java8/function/src/main/java/com/digitaldots/function/BeverRequest.java
@@ -1,6 +1,0 @@
-package com.openfaas.function;
-
-import com.openfaas.model.IRequest;
-
-public interface BeverRequest extends IRequest {
-}
