--- template/digitaldots-java8/function/src/main/java/com/digitaldots/function/BeverResponse.java
@@ -1,7 +1,0 @@
-package com.openfaas.function;
-
-import com.openfaas.model.IResponse;
-import com.openfaas.model.Response;
-
-public class BeverResponse extends Response implements IResponse {
-}
